purge
test

+++ PurgeUser.ps1
@@ -18,7 +18,7 @@ Script de nettoyage profils utilisateurs.
 		[boolean]$debugMode
 
 .EXAMPLE
-./scriptDeleteAdmin.ps1 -ToggleFolderCount 40 -Threshold 200 -MaxFolderLeft 40 -debugMode $true
+		./PurgeUser.ps1 -ToggleFolderCount 40 -Threshold 200 -MaxFolderLeft 40 -debugMode $true
 
 .NOTES
 		nbr de dossiers dans /users pour activer le déclenchement
@@ -54,11 +54,38 @@ Param(
    )
 	
 	$LogFile = "c:\temp\_logs\PurgeUser.log"
+	#$Maillinglist =@("s.monchatre@probtp.com";"n.bazillou@probtp.com";"c.castello@probtp.com";"j.meredieu@probtp.com";"p.clignac@probtp.com";"o.woillot@probtp.com";"y.viano@probtp.com";"j.degeorges@probtp.com")
+	$Maillinglist =@("d.mousel@probtp.com";"s.monchatre@probtp.com";"c.castello@probtp.com")
+
+	
+	## SENDMAIL: FONCTION D'ENVOI DE MAIL A LA MAILLINGLIST
+		function sendMail($s, $to, $p) {
+			$smtpServer = "smtp.probtp"
+			$smtpFrom = "PurgeUserScript@probtp.com"
+			$messageSubject = $s[0]
+			Foreach ($dest in $Maillinglist){		
+			$smtpTo = $dest
+			$message = New-Object System.Net.Mail.MailMessage $smtpfrom, $smtpto
+			$message.Subject = $messageSubject
+			$message.IsBodyHTML = $false
+			$message.Body = $s[1]
+			$message.Priority = $p
+			$smtp = New-Object Net.Mail.SmtpClient($smtpServer)
+			$smtp.Send($message)
+		}
+	}	
+	
+	
 	#Sécurité variable d'environement
 	if(!(Test-Path Env:\_PC_FONCTION_ACCUEIL_TEL))
 	{
 		 write-warning "Alerte : type de poste invalide" 
 		 add-content $LogFile -value "Alerte : type de poste invalide"
+		 $poste = $env:computername
+		 $subject = "PurgeUser.ps1 Alerte :  $poste type de poste invalide"
+		 $body = "type de poste invalide"
+		 $email =@($subject,$body)
+		 sendMail -s $email -to $Maillinglist -p "High"	
 		 exit 666
 	}
 	
@@ -68,12 +95,12 @@ Param(
 	$LogFile = "c:\temp\_logs\PurgeUser.log"
 	$date = get-date -format F
 	$WhiteUsersList = @("Administrateur","netshield","service_bmc_fpac","pb11997",$User,"Public")
+
 	add-content $LogFile -value "#################################################"
 	add-content $LogFile -value "$date : début nettoyage profils utilisateurs"
 	add-content $LogFile -value "`r`n"
 	add-content $LogFile -value "Param ToggleFolderCount = $ToggleFolderCount "
 	add-content $LogFile -value "Param Treshold = $Threshold "
-	add-content $LogFile -value "Param MaxFolderLeft = $MaxFolderLeft"
 	add-content $LogFile -value "Param MaxFolderLeft = $MaxFolderLeft"
 	add-content $LogFile -value "Param debug = $debugMode"
 	
@@ -92,51 +119,67 @@ Param(
 	add-content $LogFile -value "`r`n"
 	
 	#clean folder without key	
+	$countTrashFolder =0;
+	$TrashFolderLog = "";
 	write-warning "clean dossier sans clé " 
 	Get-ChildItem "$($env:SystemDrive)\Users" | ? {$WhiteUsersList -notcontains $_.Name} | % {
 	$TestPath = $_.FullName ; 
+
 	$key = Get-ChildItem "HKLM:\Software\Microsoft\Windows NT\CurrentVersion\ProfileList" | ? {$_.GetValue('ProfileImagePath') -eq "$TestPath"} | Select-Object -first 1
 	if(!$key)
 	{
-				write-host "key not found for folder $TestPath"
+	write-host " $TestPath pas de clé!"
+		$countTrashFolder++
+		
+			
 				if($debugMode -eq $false)
 				{
-						cmd /c rmdir $TestPath  /s /q
-						cmd /c rmdir $TestPath  /s /q
+						cmd /c rmdir "$TestPath"  /s /q
+						#cmd /c rmdir "$TestPath"  /s /q
 						if(!(test-path  $TestPath))
 						{
-						add-content $LogFile -value "TRASH DOSSIER -> Suppression du dossier user $TestPath réussie"
-						write-host "TRASH DOSSIER ->  Suppression du dossier user $TestPath réussie"
+						$TrashFolderLog += "TRASH DOSSIER -> Suppression du dossier user $TestPath réussie `r`n"					
 						}
 						else
 						{
-						add-content $LogFile -value "DOSSIER -> Suppression du dossier user $TestPath échoué "
-						write-host "TRASH DOSSIER -> Suppression du dossier user $TestPath échoué"
+							$TrashFolderLog += "TRASH DOSSIER -> Suppression du dossier user $TestPath échoué `r`n"						
 						}
 				}
 				else
 				{
-						add-content $LogFile -value "DEBUG TRASH DOSSIER ->  Suppression du dossier user $TestPath réussie"
-						write-host "DEBUG TRASH DOSSIER ->  Suppression du dossier user $TestPath réussie"
+							$TrashFolderLog += "DEBUG TRASH DOSSIER ->  Suppression du dossier user $TestPath réussie `r`n"
 				}
 	}
 	#$key.GetValue('ProfileImagePath')	
 	}
+	write-host $TrashFolderLog
+	add-content $LogFile -value "suppression des dossiers sans clés : $countTrashFolder"
+	add-content $LogFile -value $TrashFolderLog
 	
 	#clean key without folder
+	$countTrashKey=0;
+	$TrashKeyLog = "";
 	write-warning "clean clé sans dossier " 
 	$ProfileList = Get-ChildItem "HKLM:\Software\Microsoft\Windows NT\CurrentVersion\ProfileList" | ? {$_.GetValue('ProfileImagePath')  -notmatch 'administrator|Ctx_StreamingSvc|NetworkService|Localservice|systemprofile' -AND $_.GetValue('ProfileImagePath') -Match 'C:\\Users\\'} 
 	$ProfileList | %{
+	
 	$currentPath = $_.GetValue('ProfileImagePath')
 	if (-Not (test-path $currentPath))
-	{ 					fa
+	{ 					
 		write-host "folder not found $currentPath"
-		add-content $LogFile -value "REGISTRE TRASH KEY -> Suppression de la clé utilisateur where ProfileImagePath = $currentPath"
-						
+		$countTrashKey++
+		$TrashKeyLog += "REGISTRE TRASH KEY -> Suppression de la clés utilisateur where ProfileImagePath = $currentPath `r`n"
+		if($debugMode -eq $false)
+		{			
 		$_ | Remove-Item	
+		
+		
+		}
 	}
 
 	}
+		add-content $LogFile -value "suppression des clés sans dossier : $countTrashKey"
+		add-content $LogFile -value $TrashKeyLog
 	
 	# sélection des dossiers hors whiteList
 	$UserProfileFolders = Get-ChildItem "$($env:SystemDrive)\Users" | ? {$WhiteUsersList -notcontains $_.Name} | Select Name,FullName,LastAccessTime
@@ -148,7 +191,7 @@ Param(
 	if(test-path "$((Get-Item $env:USERPROFILE).Parent.FullName)\$name\NTUSER.DAT")
 	{	
 		$n = Get-Item "$((Get-Item $env:USERPROFILE).Parent.FullName)\$name\NTUSER.DAT" -force				
-		if(  ( (Get-Date)   - $n.LastWriteTime ).Days -ge $Threshold)
+		if(((Get-Date)   - $n.LastWriteTime ).Days -ge $Threshold)
 		{
 			write-host  $((Get-Date)   - $n.LastWriteTime ).Days" greater or equal $Threshold"
 		
@@ -174,31 +217,36 @@ Param(
 	}
 	
 	}
+	
+	
+	add-content $LogFile -value  " "
+	add-content $LogFile -value  "---------------------SUPPRESSION DES PROFILS VALIDES-----------------------------------"
+	$cf = $FinalCollection.Count
+	add-content $LogFile -value  "Collection finale count: $cf"
 	$FinalCollection  | sort-object -property NTUSERLastWriteTime | FT
 	#delete
 	#collection triée par date DESC (Delete du + ancien au + recent) 
+	$CountUserLeft = -1;
 	$FinalCollection  | sort-object -property NTUSERLastWriteTime  | % {
 
-
-		
+	$CountUserLeft++
+	$index = $FinalCollection.Length - $CountUserLeft
+	$index
 	
 		$PB = Split-Path $_.FullName -Leaf
   		$path = $_.FullName
-		add-content $LogFile -value  "------------------------------------------------"
-		add-content $LogFile -value  "début suppression $PB `r`n"
+	
 		
-		#check nbr de dossier restant si < $MaxFolderLeft  exit script
-		$CountUserProfileFolders = Get-ChildItem "$($env:SystemDrive)\Users" 
-		$c = $CountUserProfileFolders.Count ;
-		add-content $LogFile -value "Nombre de dossiers : $c dossiers" ; 
-		if($CountUserProfileFolders.Count -lt $MaxFolderLeft){
-		write-host "nombre de dossiers inferieur au $treshold exit" ;
+		if($index -le $MaxFolderLeft){
+		write-host "nombre de dossiers inferieur au MaxLeftFolder exit" ;
 		$regCount = (Get-ChildItem "HKLM:\Software\Microsoft\Windows NT\CurrentVersion\ProfileList").Count
 		add-content $LogFile -value "Nombre de clé de registre user : $regCount clé" ; 
 		
 		add-content $LogFile -value "Nombre de dossiers users inferieur à la valeur definie $MaxFolderLeft exit!" ; 
 		exit 0
 		}
+		add-content $LogFile -value  "`r`n début suppression $PB "
+		add-content $LogFile -value "Index de suppression dans la collection : $index dossiers" ; 
 		 <#		if($debugMode -eq $false)
 			{
 				Try
@@ -255,23 +303,24 @@ Param(
 				}
 			
 		
-			try
-			{
+			
 				if($debugMode -eq $false)
 				{
 					Get-ChildItem "HKLM:\Software\Microsoft\Windows NT\CurrentVersion\ProfileList" | ? {$_.GetValue('ProfileImagePath') -Match "$PB" -and $_.GetValue('ProfileImagePath') -Match $_.FullName} | Select-Object -first 1 | Remove-Item
 				
 				
 				$test = Get-ChildItem "HKLM:\Software\Microsoft\Windows NT\CurrentVersion\ProfileList" | ? {$_.GetValue('ProfileImagePath') -Match "$PB" -and $_.GetValue('ProfileImagePath') -Match $_.FullName}
-				if($test.Count -eq 0)
-						{										
-							write-host "REGISTRE -> User: $PB - Suppression clé de registre $UserKey réussie"
-							add-content $LogFile -value " REGISTRE -> User: $PB - Suppression clé de registre $UserKey réussie"
+				if($test)
+						{	
+							write-host "REGISTRE -> User: $PB - Suppression clé de registre $UserKey échoué"
+							add-content $LogFile -value " REGISTRE -> User: $PB - Suppression clé de registre $UserKey échoué"									
+							
 						}
 						else
-						{
-							write-host "REGISTRE -> User: $PB - Suppression clé de registre $UserKey échoué"
-							add-content $LogFile -value " REGISTRE -> User: $PB - Suppression clé de registre $UserKey échoué"
+						{	
+							write-host "REGISTRE -> User: $PB - Suppression clé de registre $UserKey réussie"
+							add-content $LogFile -value " REGISTRE -> User: $PB - Suppression clé de registre $UserKey réussie"
+						
 						}
 				}
 				else
@@ -280,11 +329,6 @@ Param(
 					add-content $LogFile -value "DEBUG REGISTRE -> User: $PB - Suppression clé de registre $UserKey réussie"
 				}
 				
-			}
-			catch
-			{			
-				add-content $LogFile -value "REGISTRE -> User: $PB - Suppression clé de registre  échoué"	
-				write-host "REGISTRE -> User: $PB - Suppression clé de registre  échoué"
-			}  
+			
 				
 	}
